Source `~/.ssh/rc' on client connection
  We need to customize the environment when we connect, but there is no
easy way to do it. The "normal" `~/.ssh/rc' is searched in the real home
directory, and we want our customized environment to be available in
non-interactive mode and different shells. Our solution is to always
source `~/.ssh/rc' in the fake home directory when it exists, before we
execute ssh original command.